requirements.txt
numpy
tensorflow
keras-tuner